Almost finished labday 8

+++ Node2/Drivers/motor_regulator.h
@@ -6,7 +6,7 @@ typedef struct {
 } motor_position_t;
 
 void motor_regulator_init();
-void update_motor_pos(int32_t u, float start, float end);
+void update_motor_pos(int32_t u, int32_t start, int32_t end);
 void read_encoder(motor_position_t* mot_pos);
 void control_motor(motor_position_t* mot_pos, pad_t* pad);
